% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-endpoint.R
\name{oauth_endpoint_azure}
\alias{oauth_endpoint_azure}
\title{Describe an OAuth2.0 endpoint for Azure.}
\usage{
oauth_endpoint_azure(tenant_id)
}
\arguments{
\item{tenant_id}{character}
}
\value{
endpoint object returned by \code{\link[httr:oauth_enpoint]{httr::oauth_enpoint()}}
}
\description{
This endpoint is designed for use with Azure Datalake Store; however,
it may work for other Azure services.
}
\details{
This code is based on:
\itemize{
\item \href{https://docs.microsoft.com/en-us/azure/data-lake-store/data-lake-store-get-started-rest-api#how-do-i-authenticate-using-azure-active-directory}{Azure Datalake Store documentation}
\item \href{https://github.com/hadley/httr/blob/master/demo/oauth2-azure.r}{httr demo for Azure}
}
}
\examples{
  oauth_endpoint_azure("aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee")
}
