% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-token.R
\name{oauth_service_token_azure}
\alias{oauth_service_token_azure}
\title{Generate OAuth2.0 token for Azure service account.}
\usage{
oauth_service_token_azure(tenant_id, application_id, application_secret,
  resource = "https://management.core.windows.net/")
}
\arguments{
\item{tenant_id}{character, tenant id issued by Azure}

\item{application_id}{character, application id issued by Azure}

\item{application_secret}{character, key issued by the account}

\item{resource}{character, resource requested}
}
\value{
A \code{TokenServiceAccountAzure} reference class (RC) object.
}
\description{
As described in \code{\link[httr:oauth_service_token]{httr::oauth_service_token()}}, service accounts
provide a way of using OAuth2 without user intervention.
}
\details{
At some point in the near future, describe how to make a Azure web-app.
}
\examples{
\dontrun{
  oauth_service_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_secret = "foofoofoofoofoofoofoofoofooooo"
  )
}
}
