% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-token.R
\name{oauth_token_azure}
\alias{oauth_token_azure}
\title{Generate an OAuth2.0 token for a native Azure app.}
\usage{
oauth_token_azure(tenant_id, application_id, name = NULL,
  user_params = list(resource = "https://management.core.windows.net/"),
  use_oob = FALSE, ...)
}
\arguments{
\item{tenant_id}{character, tenant id issued by Azure}

\item{application_id}{character, application id issued by Azure}

\item{name}{character, name of application (optional)}

\item{user_params}{named list, see \code{\link[httr:oauth2.0_token]{httr::oauth2.0_token()}}}

\item{use_oob}{logical, see \code{\link[httr:oauth2.0_token]{httr::oauth2.0_token()}}}

\item{...}{other arguments passed to \code{\link[httr:oauth2.0_token]{httr::oauth2.0_token()}}}
}
\value{
A \code{Token2.0} reference class (RC) object, returned by \code{\link[httr:oauth2.0_token]{httr::oauth2.0_token()}}.
}
\description{
This is a wrapper to \code{\link[httr:oauth2.0_token]{httr::oauth2.0_token()}}, dedicated to Azure OAuth access.
}
\details{
The \code{tenant_id} is used to build the endpoint using \code{\link[=oauth_endpoint_azure]{oauth_endpoint_azure()}}.
The \code{application_id} and (optional) \code{name} are used to build the
app using \code{\link[httr:oauth_app]{httr::oauth_app()}}.

On the Azure side, can build the app from the Azure Portal:
\enumerate{
\item Go to \strong{Active Directory} > \strong{App registrations} >
\strong{New application registration}.
\item For "Application type", choose "Native".
\item For "Sign-on URL", enter \code{http://localhost:1410}.
\item Compose this link using \code{tenant_id} and \code{application_id},
then send to your Azure Tenant adminstrator (maybe that's you)
to activate the permissions:
\code{https://login.microsoftonline.com/{tenant_id}/adminconsent?client_id={application_id}&state=12345&redirect_uri=http://localhost/}
}

For more information on "activating" your app, please see \href{https://docs.microsoft.com/en-us/azure/active-directory/develop/active-directory-v2-scopes}{this Azure documentation}.
}
\examples{
\dontrun{
  oauth_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj",
    name = "test-app"
  )
}
}
