% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-token.R
\docType{class}
\name{Token-class}
\alias{Token-class}
\alias{TokenServiceAccountAzure}
\title{OAuth token objects for Azure service app.}
\format{An R6 class object.}
\usage{
TokenServiceAccountAzure
}
\description{
OAuth token objects for Azure service app.
}
\seealso{
\code{\link[httr:Token-class]{httr::Token}}
}
\keyword{internal}
